requirements.txt
fastapi
uvicorn
jinja2
pydantic
torch
wikipedia
beautifulsoup4
tqdm
python-multipart
python-dotenv
google-cloud-storage
google-auth